requirements.txt
discord.py-self
colorama
python